Refactor command to helper, add tests
