+++ usaspending_api/etl/management/commands/load_usaspending_assistance.py
@@ -6,7 +6,7 @@ from django.core.management.base import BaseCommand
 from usaspending_api.awards.models import Award, AWARD_TYPES, Transaction, TransactionAssistance
 from usaspending_api.etl.award_helpers import update_awards, update_contract_awards
 from usaspending_api.etl.csv_data_reader import CsvDataReader
-from usaspending_api.etl.commands.update_description_fields import update_model_description_fields
+from usaspending_api.etl.helpers import update_model_description_fields
 import usaspending_api.etl.helpers as h
 from usaspending_api.references.models import Agency, LegalEntity
 from usaspending_api.submissions.models import SubmissionAttributes
